Fixed: page look and sorting
* fixed sorting for GB and MB values
* adjusted quotes style in js
* used double instead of int for CPU load
* graph made linear

+++ Dashboard/OverviewPageModel.cs
@@ -9,7 +9,7 @@ namespace Hangfire.Heartbeat.Dashboard
         public string Name { get; set; }
         public string ProcessId { get; set; }
         public string ProcessName { get; set; }
-        public int CpuUsagePercentage { get; set; }
+        public double CpuUsagePercentage { get; set; }
         public long WorkingMemorySet { get; set; }
         public long Timestamp { get; set; }
     }
